Remove basic auth from datanode rest api (jwt only now)
